+++ Simulation/Simulateur/src/util/Echeancier.java
@@ -8,7 +8,7 @@ import java.util.*;
 public class Echeancier {
     private static List<Pair> evenements = new ArrayList <> ();
 
-    public static void ajouterEvenement(Evenement evenement, float date) {
+    public static void ajouterEvenement(Evenement evenement, double date) {
         evenements.add(new Pair(evenement, date));
     }
